Obstaculos finalizado
- Mejorado el codigo del bonus
- Agregado de animaciones y colisiones de Obstaculos
- Mejorado el posicionamiento
- Borrado de otras "basuras"

+++ source/Reg.hx
@@ -36,70 +36,33 @@ class Reg
 			}
 		}
 	}
-	static public function ColisionAntiBala(anti:Array<AntiBala>,personaje:Personaje):Void{
-		for (i in 0...anti.length){
-			if (FlxG.overlap(personaje, anti[i])){
-				if (vidas > 0){
-				anti[i].Posicionar();
-				vidas--;
-				trace("te quedan " + vidas + " vidas");
-				} 
-				else{
-				personaje.kill();
-				trace("has muerto");
+	static public function ColisionAntiBala(anti:Array<AntiBala>,personaje:Personaje=null,obstaculo:Array<Obstaculo>=null):Void{
+		if (personaje != null){
+			for (i in 0...anti.length){
+				if (FlxG.overlap(personaje, anti[i])){
+					if (vidas > 0){
+						anti[i].Posicionar();
+						vidas--;
+						trace("te quedan " + vidas + " vidas");
+					}
+					else{
+						personaje.kill();
+						trace("has muerto");
+					}
+				}
+			}
+		}
+		if (obstaculo != null){
+			for (i in 0...anti.length){
+				for (j in 0...obstaculo.length){
+					if (FlxG.overlap(anti[i], obstaculo[j])){
+						anti[i].Posicionar();
+						obstaculo[j].hacerDanio();
+					}
 				}
 			}
 		}
 	}
-	static public function posicionarEnemigos(?enemigos:Array<Enemigo>, cantEnemigos:Int):Void{
-		var colum:Int = 8;
-		var fila:Int = 24;
-		var cantCol:Int = 5;
-		for (i in 0...cantEnemigos) 
-		{
-			if (cantCol == 0) 
-			{
-				fila += 20;
-				colum = 8;
-				cantCol = 5;
-			}
-			enemigos[i] = new Enemigo(colum, fila);
-			colum += 24;
-			cantCol--;
-		}
-	}
-	static public function posicionarObstaculos(?obstaculos:Array<Obstaculo>, cantObstaculos:Int):Void{
-		var colObs:Int = 3;
-		
-		for (i in 0...cantObstaculos) {
-
-			colObs += 30;
-			obstaculos[i] = new Obstaculo(colObs, FlxG.height - 29);		
-			cantObstaculos--;
-		}
-		
-		
-	}
-	/*static public function checkChange(enemigos:Array<Enemigo>):Void {
-		var cambioDireccion:Bool = false;	
-		for (i in 0...enemigos.length) 
-		{
-			if (enemigos[i].x + enemigos[i].width > 160){
-				Reg.movementModifier = false;
-				cambioDireccion = true;
-			}
-			else if (enemigos[i].x < 8){
-				Reg.movementModifier = true;
-				cambioDireccion = true;
-			}
-		}
-		if (cambioDireccion){
-			for (j in 0...enemigos.length){
-				enemigos[j].y += 4;
-			}
-			cambioDireccion = false; 
-		}
-	}*/
 	// no se para que esta (porque no colisionarian nunca) pero la pedia el tp
 	static public function ColisionEnemigoJugador(enemigos:Array<Enemigo>,jugador:Personaje):Void{
 		for (i in 0...enemigos.length){
